+++ ghc-tags-core/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.1.0.0 -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.
+* Normalise 'tagFilePath' for tags which are returned by the parsers.